% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corr_make}
\alias{corr_make}
\title{Construct a correlation matrix from endpoint-index triplets}
\usage{
corr_make(num_endpoints, values = NULL)
}
\description{
Construct a correlation matrix from endpoint-index triplets
}
