% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate_from_prob}
\alias{rate_from_prob}
\title{Solve exponential rate parameters from target observed event probabilities}
\usage{
rate_from_prob(
  target_prob,
  mode = c("simple", "admin", "semi-competing"),
  event_rate = NULL,
  admin_time = NULL,
  fatal_event_rate = NULL,
  fatal_censor_rate = NULL,
  nonfatal_event_rate = NULL
)
}
\description{
Solve exponential rate parameters from target observed event probabilities
}
