% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-package.R
\docType{package}
\name{endpoints-package}
\alias{endpoints}
\alias{endpoints-package}
\title{endpoints: Simulate Correlated Multi-Endpoint Clinical Trial Data}
\description{
Simulate correlated multi-endpoint clinical trial data using a Gaussian copula approach. This package provides tools to generate realistic patient-level datasets with configurable treatment effects and covariance structures. It supports multiple endpoint types (continuous, binary, count, and time-to-event) across multiple treatment arms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Boehringer-Ingelheim/endpoints}
  \item \url{https://boehringer-ingelheim.github.io/endpoints}
  \item \url{https://boehringer-ingelheim.github.io/endpoints/}
  \item Report bugs at \url{https://github.com/Boehringer-Ingelheim/endpoints/issues}
}

}
\author{
\strong{Maintainer}: Colin Longhurst \email{colin.longhurst@boehringer-ingelheim.com}

Authors:
\itemize{
  \item Ming Yang \email{ming.yang@boehringer-ingelheim.com}
  \item Mitchell Thomann \email{mitchell.thomann@boehringer-ingelheim.com}
}

Other contributors:
\itemize{
  \item Boehringer Ingelheim Pharma GmbH & Co. KG [copyright holder, funder]
}

}
